Expose formatting util functions.
